Solution_4

+++ app/main.py
@@ -2,16 +2,14 @@ import os
 
 
 def move_file(command: str) -> None:
-    split_command = command.split()
-    mv, source_file, destination_file = split_command
+    mv, source_file, destination_file = command.split()
     path = destination_file[0:destination_file.rfind("/")]
-    if mv == "mv":
-        if "/" in command:
-            if not os.path.exists(path):
-                os.makedirs(path)
-            with (open(source_file, "r") as file_in,
-                  open(destination_file, "w") as file_out):
-                file_out.write(file_in.read())
-            os.remove(source_file)
-        else:
-            os.rename(source_file, destination_file)
+    if mv == "mv" and "/" in command:
+        if not os.path.exists(path):
+            os.makedirs(path)
+        with (open(source_file, "r") as file_in,
+              open(destination_file, "w") as file_out):
+            file_out.write(file_in.read())
+        os.remove(source_file)
+    else:
+        os.rename(source_file, destination_file)
